refactor(schema): split PostgreSQL catalog reader
